This branch unifies two existing implementations of the calculation of struct sizes
and field offsets, and ensures (using assertions)
that they both produce the same results.

We can then remove one of the implementations
(the one from TypeHandlerWithPointerAliasing).
When this is done, #362 is finally fixed.


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26177 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/SMGTransferRelation.java
@@ -1236,7 +1236,7 @@ public class SMGTransferRelation
             offset += machineModel.getSizeofCharInBits() - overByte;
           }
           @SuppressWarnings("deprecation") // replace with machineModel.getAllFieldOffsetsInBits
-          int padding =
+          long padding =
               machineModel.getPadding(offset / machineModel.getSizeofCharInBits(), memberType);
           offset += padding * machineModel.getSizeofCharInBits();
         }
